+++ BlazorMonaco/Bridge/TextModel.cs
@@ -5,6 +5,6 @@ namespace BlazorMonaco.Bridge
     public class TextModel
     {
         public string Id { get; set; }
-        public Uri Uri { get; set; }
+        public string Uri { get; set; }
     }
 }